refactor: context, provider 분리
